Conflicts:
	src/main/java/com/ssl/jv/gip/negocio/ejb/MaestrosEJB.java
	src/main/java/com/ssl/jv/gip/negocio/ejb/MaestrosEJBLocal.java

+++ src/main/java/com/ssl/jv/gip/negocio/dao/IGenericDAO.java
@@ -11,7 +11,7 @@ public interface IGenericDAO<T>  {
 	
 	public void delete(T pojo);
 	
-	public T findByPK(Integer id);
+	public T findByPK(Long id);
 	
 	public List<?> findAll();
